+++ tests/test_exchange_universe.py
@@ -1,5 +1,7 @@
+import pytest
 from tradingstrategy.chain import ChainId
 from tradingstrategy.client import Client
+from tradingstrategy.exchange import ExchangeNotFoundError
 
 
 def test_look_up_by_factory(persistent_test_client: Client):
@@ -24,9 +26,12 @@ def test_look_up_by_chain_and_name(persistent_test_client: Client):
     assert quickswap
     assert quickswap.exchange_slug == "quickswap"
 
-    # there should be only 1 quickswap on Polygon
-    nonexist = exchange_universe.get_by_chain_and_name(ChainId.ethereum, "Quickswap")
-    assert nonexist is None
+    with pytest.raises(ExchangeNotFoundError) as excinfo:
+        # there should be only 1 quickswap on Polygon
+        nonexist = exchange_universe.get_by_chain_and_name(ChainId.ethereum, "Quickswap")
+        assert nonexist is None
+    
+    assert excinfo.value.args[0] == 'The trading universe does not contain data on chain ethereum for exchange_name quickswap. This might be a problem in your data loading and filtering. \n                \n    Use tradingstrategy.ai website to explore DEXs.\n    \n    Here is a list of DEXes: https://tradingstrategy.ai/trading-view/exchanges\n    \n    For any further questions join our Discord: https://tradingstrategy.ai/community'
 
 
 
@@ -41,6 +46,9 @@ def test_look_up_by_chain_and_slug(persistent_test_client: Client):
     assert quickswap
     assert quickswap.name == "Quickswap"
 
-    # there should be only 1 quickswap on Polygon
-    nonexist = exchange_universe.get_by_chain_and_slug(ChainId.ethereum, "quickswap")
-    assert nonexist is None
+    with pytest.raises(ExchangeNotFoundError) as excinfo:
+        # there should be only 1 quickswap on Polygon
+        nonexist = exchange_universe.get_by_chain_and_slug(ChainId.ethereum, "quickswap")
+        assert nonexist is None
+    
+    assert excinfo.value.args[0] == 'The trading universe does not contain data on chain ethereum for exchange_slug quickswap. This might be a problem in your data loading and filtering. \n                \n    Use tradingstrategy.ai website to explore DEXs.\n    \n    Here is a list of DEXes: https://tradingstrategy.ai/trading-view/exchanges\n    \n    For any further questions join our Discord: https://tradingstrategy.ai/community'