+++ tests/test_pair_universe.py
@@ -5,8 +5,8 @@ import pytest
 
 from tradingstrategy.chain import ChainId
 from tradingstrategy.client import Client
-from tradingstrategy.exchange import ExchangeType
-from tradingstrategy.pair import PandasPairUniverse, resolve_pairs_based_on_ticker, NoPairFound, DEXPair, generate_address_columns
+from tradingstrategy.exchange import ExchangeType, ExchangeNotFoundError
+from tradingstrategy.pair import PandasPairUniverse, resolve_pairs_based_on_ticker, PairNotFoundError, DEXPair, generate_address_columns
 
 
 @pytest.fixture
@@ -179,8 +179,10 @@ def test_resolve_based_on_human_description(persistent_test_client):
     assert bnb_busd.buy_volume_30d > 1_000_000
 
     desc = (ChainId.bsc, "pancakeswap-v2", "MIKKO", "BUSD")
-    with pytest.raises(NoPairFound):
+    with pytest.raises(PairNotFoundError) as excinfo:
         pair_universe.get_pair_by_human_description(exchange_universe, desc)
+    
+    assert excinfo.value.args[0] == 'Exchange pancakeswap-v2 does not have a pair MIKKO-BUSD with fee tier None. This might be a problem in your data loading and filtering. \n                \n    Use tradingstrategy.ai website to explore pairs. Once on a pair page, click on the `Copy Python identifier` button to get the correct pair information to use in your strategy.\n    \n    Here is a list of DEXes: https://tradingstrategy.ai/trading-view/exchanges\n    \n    Here is advanced search: https://tradingstrategy.ai/search?q=&sortBy=liquidity%3Adesc&filters=%7B%22pool_swap_fee%22%3A%5B%5D%2C%22price_change_24h%22%3A%5B%5D%2C%22liquidity%22%3A%5B%5D%2C%22volume_24h%22%3A%5B%5D%2C%22type%22%3A%5B%5D%2C%22blockchain%22%3A%5B%5D%2C%22exchange%22%3A%5B%5D%7D\n    \n    For any further questions join our Discord: https://tradingstrategy.ai/community'
 
 
 def test_get_pair(persistent_test_client):
@@ -225,6 +227,11 @@ def test_fee_tier_uniswap_v2(persistent_test_client):
     assert pair.base_token_symbol == "EUL"
     assert pair.quote_token_symbol == "WETH"
     assert pair.fee == 30
+
+    with pytest.raises(ExchangeNotFoundError) as excinfo:
+        exchange = exchange_universe.get_by_chain_and_slug(ChainId.ethereum, "alex-v2")
+
+    assert excinfo.value.args[0] == 'The trading universe does not contain data on chain ethereum for exchange_slug alex-v2. This might be a problem in your data loading and filtering. \n                \n    Use tradingstrategy.ai website to explore DEXs.\n    \n    Here is a list of DEXes: https://tradingstrategy.ai/trading-view/exchanges\n    \n    For any further questions join our Discord: https://tradingstrategy.ai/community'
 
 
 def test_fee_tier_uniswap_v3(persistent_test_client):