+++ tradingstrategy/pair.py
@@ -27,20 +27,42 @@ from numpy import isnan
 
 from tradingstrategy.chain import ChainId
 from tradingstrategy.token import Token
-from tradingstrategy.exchange import ExchangeUniverse, Exchange, ExchangeType
+from tradingstrategy.exchange import ExchangeUniverse, Exchange, ExchangeType, ExchangeNotFoundError
 from tradingstrategy.stablecoin import ALL_STABLECOIN_LIKE
 from tradingstrategy.types import NonChecksummedAddress, BlockNumber, UNIXTimestamp, BasisPoint, PrimaryKey, Percent, \
     USDollarAmount
 from tradingstrategy.utils.columnar import iterate_columnar_dicts
 from tradingstrategy.utils.schema import create_pyarrow_schema_for_dataclass, create_columnar_work_buffer, \
     append_to_columnar_work_buffer
+from tradingstrategy.exceptions import DataNotFoundError
 
 
 logger = logging.getLogger(__name__)
 
 
-class NoPairFound(Exception):
+class PairNotFoundError(DataNotFoundError):
     """No trading pair found matching the given criteria."""
+
+    advanced_search_url = "https://tradingstrategy.ai/search?q=&sortBy=liquidity%3Adesc&filters=%7B%22pool_swap_fee%22%3A%5B%5D%2C%22price_change_24h%22%3A%5B%5D%2C%22liquidity%22%3A%5B%5D%2C%22volume_24h%22%3A%5B%5D%2C%22type%22%3A%5B%5D%2C%22blockchain%22%3A%5B%5D%2C%22exchange%22%3A%5B%5D%7D"
+
+    template = f"""This might be a problem in your data loading and filtering. 
+                
+    Use tradingstrategy.ai website to explore pairs. Once on a pair page, click on the `Copy Python identifier` button to get the correct pair information to use in your strategy.
+    
+    Here is a list of DEXes: https://tradingstrategy.ai/trading-view/exchanges
+    
+    Here is advanced search: {advanced_search_url}
+    
+    For any further questions join our Discord: https://tradingstrategy.ai/community"""
+
+    def __init__(self, base_token: str, quote_token: str, fee_tier: Optional[Percent] = None, exchange_slug: Optional[str] = None):
+
+        if exchange_slug:
+            message = f"Exchange {exchange_slug} does not have a pair {base_token}-{quote_token} with fee tier {fee_tier}. {self.template}"
+        else:
+            message = f"No pair {base_token}-{quote_token} with fee tier {fee_tier} found. {self.template}"
+
+        super().__init__(message)
 
 
 class DuplicatePair(Exception):
@@ -886,7 +908,7 @@ class PandasPairUniverse:
 
             Highest volume trading pair if multiple matches.
 
-        :raise NoPairFound:
+        :raise PairNotFoundError:
             In the case input data cannot be resolved.
         """
 
@@ -984,7 +1006,7 @@ class PandasPairUniverse:
 
             Highest volume trading pair if multiple matches.
 
-        :raise NoPairFound:
+        :raise PairNotFoundError:
             In the case input data cannot be resolved.
         """
 
@@ -1013,10 +1035,7 @@ class PandasPairUniverse:
             else:
                 exchange_message = ""
 
-            raise NoPairFound(f"The trading universe does not contain data for the exchange {exchange_slug} on chain {chain_id.name}.\n"
-                              f"Did you construct the trading universe correctly?\n"
-                              f"We have data for {exchange_universe.get_exchange_count()} exchange.\n"
-                              f"{exchange_message}")
+            raise ExchangeNotFoundError(chain_id.name, exchange_slug=exchange_slug, optional_extra_message=exchange_message)
 
         pair = self.get_one_pair_from_pandas_universe(
             exchange.exchange_id,
@@ -1027,10 +1046,7 @@ class PandasPairUniverse:
         )
 
         if pair is None:
-            raise NoPairFound(f"Exchange {exchange_slug} does not have a pair {base_token}-{quote_token} with fee tier {fee_tier}.\n"
-                              f"This might be a problem in your data loadng and filtering.\n"
-                              f"Did you construct the trading universe correctly?\n"
-                              f"Did you set the fee tier correctly?")
+            raise PairNotFoundError(base_token, quote_token, fee_tier, exchange_slug)
 
         return pair
 
@@ -1099,7 +1115,7 @@ class PandasPairUniverse:
         :raise DuplicatePair:
             Multiple pairs matching the criteria
 
-        :raise NoPairFound:
+        :raise PairNotFoundError:
             No pairs matching the criteria
         """
 
@@ -1142,7 +1158,7 @@ class PandasPairUniverse:
             pick one with the highest trade volume
 
         :raise DuplicatePair: Multiple pairs matching the criteria
-        :raise NoPairFound: No pairs matching the criteria
+        :raise PairNotFoundError: No pairs matching the criteria
         """
 
         warnings.warn('This method is deprecated. Use PandasPairUniverse.create_pair_universe() instead', DeprecationWarning, stacklevel=2)
@@ -1176,7 +1192,7 @@ class PandasPairUniverse:
                 frames.append(filtered_df)
 
             else:
-                raise NoPairFound(f"No trading pair found. Exchange:{exchange} base:{base_token_symbol} quote:{quote_token_symbol}")
+                raise PairNotFoundError(base_token_symbol, quote_token_symbol, exchange_slug=exchange.exchange_slug)
 
         return PandasPairUniverse(pd.concat(frames))
 
@@ -1297,7 +1313,7 @@ class LegacyPairUniverse:
         """
         return self.pairs[pair_id]
 
-    def get_pair_by_ticker(self, base_token, quote_token) -> Optional[DEXPair]:
+    def get_pair_by_ticker(self, base_token: str, quote_token: str) -> Optional[DEXPair]:
         """Get a trading pair by its ticker symbols.
 
         Note that this method works only very simple universes, as any given pair
@@ -1315,7 +1331,7 @@ class LegacyPairUniverse:
         if pairs:
             return pairs[0]
 
-        return None
+        raise PairNotFoundError(base_token, quote_token)
 
     def get_pair_by_ticker_by_exchange(self, exchange_id: int, base_token: str, quote_token: str) -> Optional[DEXPair]:
         """Get a trading pair by its ticker symbols.